requirements.txt
openai>=1.0.0
pandas>=1.5.0
python-dotenv>=0.19.0
tabulate>=0.9.0